fix: derive central-directory offset independent of the claimed field
A real P1 bypass in the previous round's entry-counting fix, confirmed
and reproduced:

_actual_central_directory_entry_count() seeked to the raw,
EOCD/ZIP64-record-declared cd_offset -- but that field is written
relative to the zip data's own start, not the whole file, so a
concatenated/self-extracting archive with bytes prepended before the
zip structure makes it wrong by exactly the prefix length.
zipfile.ZipFile itself never trusts this field directly either:
CPython's own _RealGetContents derives start_dir = eocd_location -
size_cd, where the claimed offset field cancels out of the formula
entirely.

Reproduced: a prefixed archive whose EOCD total_entries was patched to
1 counted zero records with the old, offset-trusting seek (silently
passing) while ZipFile itself correctly rebased and materialized all
20,001 real ZipInfo objects.

Fixed by deriving the same offset-independent position CPython uses
(record_position - cd_size, where record_position is the verified
location of whichever EOCD/ZIP64 record supplied cd_size -- never the
claimed offset field, which is now not even read).

tests/test_bundle_archive_cd_guard.py (new): the central-directory
guard and preflight-safety test classes, split out of
tests/test_bundle_archive.py purely to keep both under the ADR-061
1200-line test cap after these additions.
